requirements.txt
certifi==2019.3.9
chardet==3.0.4
entrypoints==0.3
flake8==3.7.7
idna==2.8
luma.core==1.8.3
luma.led-matrix==1.2.0
mccabe==0.6.1
Pillow==6.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
smbus2==0.2.3
urllib3==1.24.3
